Moving off priviledged ports
This is cool for running locally and folks can modify and rebuild easily enough. But, to run ssl-tester.chrisshort.net from a production server, I have to use a web server to route traffic.

This should be proxy passed through so that you actually see the cert that the go app is using.